requirements.txt
django-extensions==3.1.5
django-phonenumber-field==5.0.0
django-debug-toolbar
django-crispy-forms
python-dotnev
django-rest-framework
psycopg2-binary
django-treebeard
pillow